requirements.txt
av==14.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.6.0
evdev==1.9.1
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.4
onnxruntime==1.21.0
packaging==24.2
pillow==11.1.0
protobuf==6.30.2
pycparser==2.22
pyperclip==1.9.0
python-dotenv==1.1.0
python-xlib==0.33
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.3
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
PySide6==6.7.0
psutil==5.9.8
